requirements.txt
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0

pandas==2.3.3
numpy==2.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

matplotlib==3.10.7
cycler==0.12.1
kiwisolver==1.4.9
contourpy==1.3.3
pyparsing==3.2.5
fonttools==4.60.1
pillow==12.0.0
packaging==25.0

reportlab==4.4.5

openpyxl==3.1.5
et_xmlfile==2.0.0

gunicorn==23.0.0
pymongo
pillow
